requirements.txt
sqlalchemy>=2.0.48
psycopg2>=2.9.11
python-dotenv>=0.9.9
pandas>=2.3.3
numpy>=2.2.6